+++ extension/scripts/background.js
@@ -7,10 +7,14 @@
 import { initMenuManager } from "./background/menu-manager.js";
 import { initDownloadManager } from "./background/download-manager.js";
 import { initMessageHandler } from "./background/message-handler.js";
-import { initWebRequestInterceptor } from "./url monitor/web-request-interceptor.js";
+import { initWebRequestInterceptor } from "./background/web-request-interceptor.js";
 import { createDataStore, cleanupOldItems } from "./background/data-store.js";
 import { Logger } from "./utils/logger.js";
-import { UI_CONSTANTS, TIME_CONSTANTS, MODULE_NAMES } from "./utils/constants.js";
+import {
+  UI_CONSTANTS,
+  TIME_CONSTANTS,
+  MODULE_NAMES,
+} from "./utils/constants.js";
 
 // 創建模組特定的日誌記錄器
 const logger = Logger.createModuleLogger(MODULE_NAMES.BACKGROUND);
