+++ extension/scripts/content/message-handler.js
@@ -9,8 +9,8 @@ import {
   MESSAGE_ACTIONS,
   MODULE_NAMES,
 } from "../utils/constants.js";
-import { handleGetAudioDurationRequest } from "../url monitor/audio-analyzer.js";
-import { handleExtractBlobRequest } from "../url monitor/blob-monitor.js";
+import { handleGetAudioDurationRequest } from "../page-context/audio-analyzer.js";
+import { handleExtractBlobRequest } from "../page-context/blob-monitor.js";
 
 // 創建模組特定的日誌記錄器
 const logger = Logger.createModuleLogger(MODULE_NAMES.CONTENT_MESSAGE_HANDLER);
