+++ extension/scripts/page-context/audio-analyzer.js
