+++ extension/scripts/page-context/blob-analyzer.js
