+++ extension/scripts/page-context/blob-monitor.js
